create dinamy map, update world and struture

+++ src/utils/coordinates.rs
@@ -1,9 +1,10 @@
 pub const TILE_SIZE: f32 = 32.0;
 
-pub fn to_screen_x(x: usize, y: usize) -> f32 {
-    ((x as isize - y as isize) as f32) * TILE_SIZE / 2.0
+/// Proyección isométrica (tipo TutsPlus)
+pub fn to_screen_x(x: i32, y: i32) -> f32 {
+    (x - y) as f32 * TILE_SIZE / 2.0
 }
 
-pub fn to_screen_y(x: usize, y: usize) -> f32 {
-    ((x as isize + y as isize) as f32) * TILE_SIZE / 4.0
+pub fn to_screen_y(x: i32, y: i32) -> f32 {
+    (x + y) as f32 * TILE_SIZE / 4.0
 }
